requirements.txt
adal==1.2.2
argon2-cffi==19.2.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.23
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.48
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-redis==5.0.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
beautifulsoup4==4.8.1
blinker==1.4
boto3==1.10.27
botocore==1.13.27
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
Django==2.2.7
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-dia==0.4.1
django-extensions==2.2.8
django-model-utils==4.0.0
docutils==0.15.2
Flask==1.1.1
Flask-Navigation==0.2.0
graphviz==0.13.2
gunicorn==20.0.4
idna==2.8
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
mysql-connector==2.2.9
oauthlib==3.1.0
packaging==19.2
Pillow==7.0.0
pluggy==0.13.0
psycopg2-binary==2.8.4
py==1.8.0
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.8
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2
